// MarkEd-front/src/api/models/PersonalAdjustmentSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PersonalAdjustmentSchema = {
    student_id: number;
    userNumber: string;
    userName: string;
    group_score: number;
    group_total: number;
    adjustment_score: number;
    adjustment_reason?: (string | null);
    final_score: number;
    status: string;
};